+++ src/engine/Events.ts
@@ -8,7 +8,7 @@ import { TileMap } from './TileMap';
 import { Side } from './Collision/Side';
 import * as Input from './Input/Index';
 import { CollisionContact } from './Collision/Detection/CollisionContact';
-import { Collider } from './Collision/Shapes/Collider';
+import { Collider } from './Collision/Colliders/Collider';
 import { Entity } from './EntityComponentSystem/Entity';
 import { OnInitialize, OnPreUpdate, OnPostUpdate } from './Interfaces/LifecycleEvents';
 import { BodyComponent } from './Collision/BodyComponent';